[MIG] survey_question_type_binary: Migration to 18.0
