+++ survey_question_type_binary/README.rst
@@ -17,13 +17,13 @@ Survey binary question type
     :target: http://www.gnu.org/licenses/agpl-3.0-standalone.html
     :alt: License: AGPL-3
 .. |badge3| image:: https://img.shields.io/badge/github-OCA%2Fsurvey-lightgray.png?logo=github
-    :target: https://github.com/OCA/survey/tree/17.0/survey_question_type_binary
+    :target: https://github.com/OCA/survey/tree/18.0/survey_question_type_binary
     :alt: OCA/survey
 .. |badge4| image:: https://img.shields.io/badge/weblate-Translate%20me-F47D42.png
-    :target: https://translation.odoo-community.org/projects/survey-17-0/survey-17-0-survey_question_type_binary
+    :target: https://translation.odoo-community.org/projects/survey-18-0/survey-18-0-survey_question_type_binary
     :alt: Translate me on Weblate
 .. |badge5| image:: https://img.shields.io/badge/runboat-Try%20me-875A7B.png
-    :target: https://runboat.odoo-community.org/builds?repo=OCA/survey&target_branch=17.0
+    :target: https://runboat.odoo-community.org/builds?repo=OCA/survey&target_branch=18.0
     :alt: Try me on Runboat
 
 |badge1| |badge2| |badge3| |badge4| |badge5|
@@ -52,7 +52,7 @@ Bug Tracker
 Bugs are tracked on `GitHub Issues <https://github.com/OCA/survey/issues>`_.
 In case of trouble, please check there if your issue has already been reported.
 If you spotted it first, help us to smash it by providing a detailed and welcomed
-`feedback <https://github.com/OCA/survey/issues/new?body=module:%20survey_question_type_binary%0Aversion:%2017.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
+`feedback <https://github.com/OCA/survey/issues/new?body=module:%20survey_question_type_binary%0Aversion:%2018.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
 
 Do not contact contributors directly about support or help with technical issues.
 
@@ -82,6 +82,6 @@ OCA, or the Odoo Community Association, is a nonprofit organization whose
 mission is to support the collaborative development of Odoo features and
 promote its widespread use.
 
-This module is part of the `OCA/survey <https://github.com/OCA/survey/tree/17.0/survey_question_type_binary>`_ project on GitHub.
+This module is part of the `OCA/survey <https://github.com/OCA/survey/tree/18.0/survey_question_type_binary>`_ project on GitHub.
 
 You are welcome to contribute. To learn how please visit https://odoo-community.org/page/Contribute.
